feat(constants): add constants to @gauzy/constants package
